Move all help-related views into a dedicated 'app' …leaving all content and urls unchanged
