+++ mtp_send_money/templates/help_area/faq.html
@@ -34,7 +34,7 @@
 
       <dt>Can I send money to a prisoner using this service?</dt>
       <dd>
-        <p>Yes. The digital service (bank transfers and debit cards) is available to all <a href="{% url 'send_money:prison_list' %}">prisons in England and Wales</a>.</p>
+        <p>Yes. The digital service (bank transfers and debit cards) is available to all <a href="{% url 'prison_list' %}">prisons in England and Wales</a>.</p>
       </dd>
 
       <dt>How long will it take for the prisoner to receive the money?</dt>
