+++ mtp_send_money/templates/send_money/bank-transfer-reference.html
@@ -79,7 +79,7 @@
           </li>
         </ol>
         <p>
-          {% url 'send_money:help_bank_transfer' as help_url %}
+          {% url 'help_bank_transfer' as help_url %}
           {% url 'submit_ticket' as ticket_url %}
           {% blocktrans trimmed %}
             <a href="{{ help_url }}">Find further help</a> on how to use the service or <a href="{{ ticket_url }}">contact us</a>.
