+++ mtp_send_money/templates/send_money/contact-form.html
@@ -61,17 +61,17 @@
       </p>
       <ul class="list list-bullet">
         <li>
-          <a href="{% url 'send_money:help_delays' %}">
+          <a href="{% url 'help_delays' %}">
             {% trans 'Payment has taken longer than expected' %}
           </a>
         </li>
         <li>
-          <a href="{% url 'send_money:help_transfered' %}">
+          <a href="{% url 'help_transfered' %}">
             {% trans 'Prisoner has been moved or released' %}
           </a>
         </li>
         <li>
-          <a href="{% url 'send_money:help_bank_transfer' %}">
+          <a href="{% url 'help_bank_transfer' %}">
             {% trans 'Bank transfer was returned or didn’t reach the prisoner' %}
           </a>
         </li>
