+++ mtp_send_money/templates/send_money/prisoner-details-error-summary.html
@@ -22,7 +22,7 @@
             <ul class="list list-bullet">
               <li>{% trans 'you entered the details correctly' %}</li>
               <li>
-                {% url 'send_money:prison_list' as prison_list_url %}
+                {% url 'prison_list' as prison_list_url %}
                 {% blocktrans trimmed %}
                   you’re sending money to a <a href="{{ prison_list_url }}">public prison</a>
                 {% endblocktrans %}
